+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [NEXT] - YYYY-MM-DD
 - Add command to dump item stack render type data
 - Adjust render type heuristic: Signs, beds and shields now use item rendering
+- Fix tile rotation not being read
 
 ## [5.0.3] - 2020-11-03
 - Fix no last selection crash